resolve comment: rename disk_used to disk_usage
