+++ src/module.rs
@@ -93,7 +93,7 @@ pub const ALL_MODULES: &[&str] = &[
     "sudo",
     "swift",
     "terraform",
-    "disk_used",
+    "disk_usage",
     "time",
     "typst",
     "username",
